Fix search() for currently provisioned node
In order for Chef's search function to find the node which is currently being
provisioned, the .json file that describes it must be part of the sandbox that
gets transferred.  Thus #create_node method must be called *before* #super in
the Kitchen::Provisioner::Nodes#create_sandbox method since Kitchen::Provisioner
::Chef::CommonSandbox#create_sandbox calls `prepare(:nodes)` which just copies
the *current* content of `test/integration/nodes`.