Update IP finder to work for CentOS 7.1
The output from ifconfig on CentOS is different from the format which was
present in the existing tests which resulted in the node['ipaddress'] value
being set to '0..5'. Instead of doing string offsets to find the IP, this uses a
regex to find the first quad-octet string that looks like an IP. It's not a
perfect IP matcher, but it works for these scenarios.